% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling-data.R
\docType{data}
\name{dat_cond_logreg}
\alias{dat_cond_logreg}
\title{Dataset for Examples in cond_logreg}
\source{
Simulated data in R.
}
\description{
List containing (1) data frame with poolwise
(g, X1, X0, C1.model, C0.model, C1.match, C0.match) values, (2) list of
replicate Xtilde values for case pools, and (3) list of replicate Xtilde
values for control pools.
}
