% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling-data.R
\docType{data}
\name{dat_p_gdfa}
\alias{dat_p_gdfa}
\title{Dataset for Examples in p_gdfa}
\source{
Simulated data in R.
}
\description{
List containing (1) data frame with poolwise (g, Y, X, Xtilde) values, (2)
list with replicate Xtilde values, and (3) list with C values for members of
each pool.
}
