% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling-data.R
\docType{data}
\name{dat_p_ndfa}
\alias{dat_p_ndfa}
\title{Dataset for Examples in p_ndfa}
\source{
Simulated data in R.
}
\description{
List containing (1) data frame with poolwise (g, Y*, Y, X, Xtilde, C) values
and (2) list with replicate Xtilde values.
}
