% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_pools.R
\name{form_pools}
\alias{form_pools}
\title{Created a Pooled Dataset from a Subject-Specific One}
\usage{
form_pools(dat, pool_sizes, num_each = NULL,
  prop_each = rep(1/length(pool_sizes), length(pool_sizes)))
}
\arguments{
\item{dat}{Data frame with individual level data.}

\item{pool_sizes}{Integer vector of pool sizes ordered from largest to
smallest.}

\item{num_each}{Integer vector specifying number of pools of each size.}

\item{prop_each}{Numeric vector specifying proportion of pools of each size.}
}
\value{
Data frame.
}
\description{
Useful for simulation studies on biospecimen pooling designs.
}
