% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_dfa_xerrors.R
\name{p.dfa.xerrors}
\alias{p.dfa.xerrors}
\title{Discriminant Function Approach for Estimating Odds Ratio with Poolwise Data
and Errors in Continuous Exposure}
\usage{
p.dfa.xerrors(g, y, xtilde, c = NULL, error.type = "both", ...)
}
\arguments{
\item{g}{Numeric vector of pool sizes (number of individuals in each pool).}

\item{y}{Numeric vector of poolwise \code{Y} values (number of cases in each
pool).}

\item{xtilde}{Numeric vector (or list of numeric vectors, if some pools have
replicate measurements) of poolwise \code{Xtilde} values.}

\item{c}{Numeric matrix of poolwise \strong{\code{C}} values, if any (columns
are variables, rows are pools).}

\item{error.type}{Character string specifying the errors that \code{X} is
subject to.  Possible values are \code{"neither"} for neither processing
error nor measurement error, \code{"processing"} for processing error only,
\code{"measurement"} for measurement error only, and \code{"both"} for both.}

\item{...}{Additional arguments to pass to \code{\link{nlminb}} function for
maximizing the log-likelihood function.}
}
\value{
A list containing the following:
\enumerate{
  \item Named numeric vector with point estimates for \code{gamma_y},
  \code{sigsq}, and the covariate-adjusted log-odds ratio, and the estimated
 variance for the log-odds ratio estimate if \code{estimate.var = TRUE}.
  \item Object returned by \code{\link[stats]{nlminb}} containing information
  about likelihood maximization.
  \item Akaike information criterion (AIC).
 }
}
\description{
Implements method described in [1].
}
\references{
1. Lyles, R.H., Van Domelen, D.R., Mitchell, E.M. and
Schisterman, E.F. (2015) "A Discriminant Function Approach to Adjust for
Processing and Measurement Error When a Biomarker is Assayed in Pooled
Samples."

Acknowledgment: This material is based upon work supported by the
National Science Foundation Graduate Research Fellowship under Grant No.
DGE-0940903.
}
