% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_logreg_xerrors2.R
\name{p.logreg.xerrors2}
\alias{p.logreg.xerrors2}
\title{Poolwise Logistic Regression with Gamma-Distributed Exposure Subject to
Errors}
\usage{
p.logreg.xerrors2(g = NULL, y, xtilde, c = NULL, error.type = "both",
  diff.pe = FALSE, diff.me = FALSE, p_y1 = NULL, p_sample.y1y0 = NULL,
  integrate.tol = 1e-08, integrate.tol.start = integrate.tol,
  integrate.tol.hessian = integrate.tol, estimate.var = FALSE, ...)
}
\arguments{
\item{g}{Numeric vector of pool sizes (number of individuals in each pool).
Do not have to specify if \code{c} is specified (function can figure it out).}

\item{y}{Numeric vector with poolwise \code{Y} values, coded 0 if all members
are controls and 1 if all members are cases.}

\item{xtilde}{Numeric vector (or list of numeric vectors, if some pools have
replicate measurements) with poolwise \code{Xtilde} values.}

\item{c}{List where each element is a matrix containing the \code{\bold{C}}
values for individual members of a particular pool. Each row of the matrix
should contain the \code{\bold{C}} values for one individual. Can leave as
\code{NULL} if there are no \code{\bold{C}} variables.}

\item{error.type}{Character string specifying the errors that \code{X} is
subject to.  Possible values are \code{"neither"} for neither processing
error nor measurement error, \code{"processing"} for processing error only,
\code{"measurement"} for measurement error only, and \code{"both"} for both.}

\item{p_y1}{Optional numeric value specifying disease prevalence, allowing
for valid estimation of the intercept with case-control sampling.  If it's
easier, you can specify \code{p_sample_y1y0} instead.  Only used if
\code{offset.formula = 1}.}

\item{p_sample.y1y0}{Optional numeric vector if length 2 specifying sampling
probabilities for cases and controls, allowing for valid estimation of the
intercept with case-control sampling.  If it's easier, you can specify
\code{p_y1} instead.}

\item{integrate.tol}{Numeric value specifying the \code{tol} input to
\code{\link{adaptIntegrate}}.  Only used if \code{approx.integral = FALSE}.}

\item{integrate.tol.start}{Same as \code{integrate.tol}, but applies only to
the very first iteration of ML maximization.  The first iteration tends to
take much longer than subsequent ones, so less precise integration at the
start can speed things up.}

\item{integrate.tol.hessian}{Same as \code{integrate.tol}, but for use when
estimating the Hessian matrix only.  Sometimes more precise integration
(i.e. smaller tolerance) than used for maximizing the likelihood helps
prevent cases where the inverse Hessian is not positive definite.}

\item{estimate.var}{If \code{TRUE}, function returns variance-covariance
matrix for parameter estimates, calculated as the inverse of the estimated
Hessian matrix at the MLE's.

#' @return A list containing the following:
\enumerate{
  \item Numeric vector of parameter estimates.
  \item Variance-covariance matrix (if \code{estimate.var = TRUE}).
  \item The returned \code{\link[stats]{nlminb}} object from maximizing the
log-likelihood function.
  \item Akaike information criterion (AIC).
}}
}
\description{
Similar to \code{\link{p.logreg.xerrors}}, but for skewed exposure that takes
on only positive values. Assumes a constant-scale Gamma model for exposure
given covariates, and multiplicative lognormal processing errors and
measurement errors acting on the poolwise mean exposure. Borrows ideas from
[1-3]. Manuscript fully describing the approach is in preparation.
}
