% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_dfa_xerrors.R
\name{p_dfa_xerrors}
\alias{p_dfa_xerrors}
\title{Discriminant Function Approach for Estimating Odds Ratio with Poolwise Data
and Errors in Continuous Exposure}
\usage{
p_dfa_xerrors(g, y, xtilde, c = NULL, errors = "both", ...)
}
\arguments{
\item{g}{Numeric vector with pool sizes, i.e. number of members in each pool.}

\item{y}{Numeric vector of poolwise \code{Y} values (number of cases in each
pool).}

\item{xtilde}{Numeric vector (or list of numeric vectors, if some pools have
replicates) with \code{Xtilde} values.}

\item{c}{Numeric matrix with poolwise \strong{\code{C}} values (if any), with
one row for each pool. Can be a vector if there is only 1 covariate.}

\item{errors}{Character string specifying the errors that \code{X} is subject
to. Choices are \code{"neither"}, \code{"processing"} for processing error
only, \code{"measurement"} for measurement error only, and \code{"both"}.}

\item{...}{Additional arguments to pass to \code{\link[stats]{nlminb}}.}
}
\value{
List containing:
\enumerate{
  \item Named numeric vector with point estimates for \code{gamma_y},
  \code{sigsq}, and the covariate-adjusted log-odds ratio, and the estimated
 variance for the log-odds ratio estimate if \code{estimate.var = TRUE}.
  \item Object returned by \code{\link[stats]{nlminb}} containing information
  about likelihood maximization.
  \item Akaike information criterion (AIC).
 }
}
\description{
Implements method described in [1].
}
\references{
Lyles, R.H., Van Domelen, D.R., Mitchell, E.M. and Schisterman, E.F. (2015)
"A Discriminant Function Approach to Adjust for Processing and Measurement
Error When a Biomarker is Assayed in Pooled Samples."
\emph{Int. J. Environ. Res. Public Health} \strong{12}(11): 14723--14740.

Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
