% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_logreg.R
\name{p_logreg}
\alias{p_logreg}
\title{Poolwise Logistic Regression}
\usage{
p_logreg(g, y, x, method = "glm", p_y1 = NULL, p_sample.y1y0 = NULL,
  estimate.var = TRUE, ...)
}
\arguments{
\item{g}{Numeric vector with pool sizes (number of members in each pool).}

\item{y}{Numeric vector with poolwise \code{Y} values, coded 0 if all members
are controls and 1 if all members are cases.}

\item{x}{Numeric matrix with poolwise \strong{\code{X}} values, with one row
for each pool. Can be a vector if there is only 1 predictor.}

\item{method}{Character string specifying method to use for estimation.
Choices are "glm" for \code{\link[stats]{glm}} function and \code{"ml"} for
maximum likelihood.}

\item{p_y1}{Numeric value specifying disease prevalence, allowing
for valid estimation of the intercept with case-control sampling. Can specify
\code{p_sample.y1y0} instead if sampling rates are known.}

\item{p_sample.y1y0}{Numeric vector of length 2 specifying sampling
probabilities for cases and controls, allowing for valid estimation of the
intercept with case-control sampling. Can specify \code{p_y1} instead if it's
easier.}

\item{estimate.var}{Logical value for whether to return variance-covariance
matrix for parameter estimates.}

\item{...}{Additional arguments to pass to \code{\link[stats]{nlminb}}.}
}
\value{
List containing:
\enumerate{
\item Numeric vector of parameter estimates.
\item Variance-covariance matrix (if \code{estimate.var = TRUE}).
\item Fitted \code{\link[stats]{glm}} object (if \code{method = "glm"}) or
returned \code{\link[stats]{nlminb}} object (if \code{method = "ml"}).
\item Akaike information criterion (AIC).
}
}
\description{
Fit logistic regression model with variables measured in pools rather than
individual participants. Requires that each pool is comprised of all cases or
all controls.
}
\details{
\emph{Statistical setup:}

The goal is to estimate coefficients in an individual-level logistic
regression model relating a continuous outcome \code{Y} to covariates
\code{\strong{X}}:

\code{logit[P(Y_ij = 1)] = beta_0 + \bold{beta_x}^T \bold{X_ij}}

But we observe poolwise values rather than individual values. In the
\code{i}th pool, which has \code{g_i} members that are either all cases or
all controls, we observe the following variables:

\code{Y_i = 1} if all \code{Y_ij = 1} and \code{0} if all \code{Y_ij = 0,
j = 1, ..., g_i} \cr
\code{\bold{X_i} = sum_{j=1}^{g_i} \bold{X_ij}}

A corresponding poolwise model has been described by Weinberg \& Umbach
(1999, 2014) and Lyles et al. (2016):

\code{logit[P(Y_i = 1)] = g_i beta_0 + \bold{beta_x}^T \bold{X_i} + qg_i}

Here \code{qg_i} is an offset defined as:

\code{qg_i = ln(\# case pools of size g_i) / \# control pools of size g_i) +
g_i ln(n_0 / n_1) + ln[P(A|Y=1) / P(A|Y=0)]}

where \code{n_0 (n_1)} represent the number of individual controls (cases) in
the study, across all pools, and \code{P(A|Y=1)} and \code{P(A|Y=0)} are
accrual probabilities for cases and controls. With prospective sampling,
these latter probabilities are equal and so the third term in the offset
drops out. With case-control sampling, these probabilities can be specified
through the \code{p_sample.y1y0} input, or alternatively (and equivalently)
the disease prevalence can be specified through the \code{p_y1} input. If
neither input is specified, the \code{beta_0} estimate will not be valid.

Thus, with poolwise data, one can fit a no-intercept logistic regression
model with \code{g_i} included as a covariate and with an offset of
\code{qg_i}. If \code{method = "glm"}, this approach is taken.

If \code{method = "ml"}, maximum likelihood is used instead. The likelihood
contribution for each observed poolwise \code{(Y_i, \bold{X_i})} is
\code{f(Y_i, \bold{X_i})}, which is proportional to \code{f(Y_i|\bold{X_i})},
which is the Bernoulli probability mass function corresponding to
\code{Y_i|\bold{X_i} ~ Bern(p_i = [g_i beta_0* + \bold{beta_x}^T \bold{X_i} +
ln(qg_i)]^(-1))}.
}
\references{
Lyles, R.H., Mitchell, E.M., Weinberg, C.R., Umbach, D.M. and Schisterman,
E.F. (2016) "An Efficient Design Strategy for Logistic Regression Using
Outcome-Dependent Pooling of Biospecimens Prior to Assay." \emph{Biometrics}
\strong{72}: 965--975.

Weinberg, C.R. and Umbach, D.M. (1999) "Using Pooled Exposure Assessment to
Improve Efficiency in Case-Control Studies." \emph{Biometrics} \strong{55}:
718--726.

Weinberg, C.R. and Umbach, D.M. (2014) "Correction to 'Using Pooled Exposure
Assessment to Improve Efficiency in Case-Control Studies' by Clarice R.
Weinberg and David M. Umbach; 55, 718--726, September 1999."
\emph{Biometrics} \strong{70}: 1061.

Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
