% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_logreg_xerrors_gamma.R
\name{p_logreg_xerrors_gamma}
\alias{p_logreg_xerrors_gamma}
\title{Poolwise Logistic Regression with Gamma Exposure Subject to Errors}
\usage{
p_logreg_xerrors_gamma(g = NULL, y, xtilde, c = NULL, errors = "both",
  nondiff.pe = TRUE, nondiff.me = TRUE, constant.pe = TRUE, p_y1 = NULL,
  p_sample.y1y0 = NULL, integrate.tol = 1e-08,
  integrate.tol.start = integrate.tol,
  integrate.tol.hessian = integrate.tol, estimate.var = FALSE, ...)
}
\arguments{
\item{g}{Numeric vector of pool sizes (number of individuals in each pool).
Can leave as \code{NULL} if \code{c} is specified (function can figure it
out).}

\item{y}{Numeric vector with poolwise \code{Y} values, coded 0 if all members
are controls and 1 if all members are cases.}

\item{xtilde}{Numeric vector (or list of numeric vectors, if some pools have
replicates) with \code{Xtilde} values.}

\item{c}{List where each element is a numeric matrix containing the
\code{\bold{C}} values for members of a particular pool (1 row for each
member).}

\item{errors}{Character string specifying the errors that \code{X} is subject
to. Choices are \code{"neither"}, \code{"processing"} for processing error
only, \code{"measurement"} for measurement error only, and \code{"both"}.}

\item{integrate.tol}{Numeric value specifying the \code{tol} input to
\code{\link{adaptIntegrate}}.}

\item{estimate.var}{If \code{TRUE}, function returns variance-covariance
matrix for parameter estimates, calculated as the inverse of the estimated
Hessian matrix at the MLE's.

#' @return A list containing the following:
\enumerate{
  \item Numeric vector of parameter estimates.
  \item Variance-covariance matrix (if \code{estimate.var = TRUE}).
  \item The returned \code{\link[stats]{nlminb}} object from maximizing the
log-likelihood function.
  \item Akaike information criterion (AIC).
}}

\item{...}{Additional arguments to pass to \code{\link[stats]{nlminb}}.}
}
\description{
Assumes constant-scale Gamma model for exposure given covariates, and
multiplicative lognormal processing errors and measurement errors acting on
the poolwise mean exposure. Borrows ideas from Weinberg \& Umbach (1999),
Schisterman et al. (2015), and Lyles et al. (2015). Manuscript fully
describing the approach is in preparation.
}
