% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling-data.R
\docType{data}
\name{pdat_matched}
\alias{pdat_matched}
\title{Dataset with Simulated Values for Examples in cond_logreg}
\source{
Simulated data in R.
}
\description{
Includes 50 pools of size 1 and 75 pools of size 2. The error-prone
\code{Xtilde} values for case pools and control pools are stored as separate
lists called \code{\link{xtilde1_matched}} and \code{\link{xtilde0_matched}}.
}
