% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lognormal_gamma.R,
%   R/plot_lognormal_gamma_old.R
\name{plot_lognormal_gamma}
\alias{plot_lognormal_gamma}
\alias{plot_lognormal_gamma}
\title{Plot Fitted Lognormal and Gamma Regression Models for Y vs. X}
\usage{
plot_lognormal_gamma(n = 1000, beta_0 = 0, beta_c = 0.5, sigsq = 0.1,
  b = NULL)

plot_lognormal_gamma(n = 1000, beta_0 = 0, beta_c = 0.5, sigsq = 0.1,
  b = NULL)
}
\arguments{
\item{n}{Numeric value.}

\item{beta_0}{Numeric value.}

\item{beta_c}{Numeric value.}

\item{sigsq}{Numeric value.}

\item{b}{Numeric value}

\item{x}{Numeric vector.}

\item{y}{Numeric vector.}

\item{n}{Numeric value.}

\item{beta_0}{Numeric value.}

\item{beta_c}{Numeric value.}

\item{sigsq}{Numeric value.}

\item{b}{Numeric value}
}
\value{
Plot of E(X) vs. X according to each fitted model.

Plot of E(X) vs. X according to each fitted model.
}
\description{
User can specify parameters under which to generate X and Y, or specify X and
Y directly (e.g. if comparing model fits for real data).

Useful for assessing whether lognormal and Gamma models are similar for
various parameter values.
}
\details{
If \code{x} and \code{y} are not specified, data are generated as lognormal
or Gamma according to the following models. If \code{b} is not specified:

X ~ N(0, 1)
Y|X ~ Lognormal(beta_0 + beta_x X, sigsq)

If \code{b} is specified:

X ~ N(0, 1)
Y|X ~ Gamma(exp(beta_0 + beta_x X), b)

Both models are fit using maximum likelihood, and the data are plotted with
the fitted E(Y|X) vs. X curves for each model.

If \code{b} is not specified, data are generated:

C ~ N(0, 1)
X|C ~ Lognormal(beta_0 + beta_c C, sigsq)

If \code{b} is specified, data are generated:

C ~ N(0, 1)
X|C ~ Gamma(exp(beta_0 + beta_c C), b)

Both models are fit using maximum likelihood, and a plot is generated
showing E(X) vs. C for each.
}
