% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolcushion_t.R
\name{poolcushion_t}
\alias{poolcushion_t}
\title{Visualize T-test Power for Pooling Design as Function of Processing Error
Variance}
\usage{
poolcushion_t(g = 2, n = 100, d = NULL, mu1 = NULL, mu2 = NULL,
  sigsq = NULL, sigsq1 = NULL, sigsq2 = NULL, sigsq_p_range = NULL,
  sigsq_m = 0, multiplicative = FALSE, alpha = 0.05, labels = TRUE,
  ylim = NULL)
}
\arguments{
\item{g}{Numeric value specifying the pool size.}

\item{n}{Numeric value specifying the number of assays per group.}

\item{d}{Numeric value specifying true difference in group means.}

\item{mu1, mu2}{Numeric value specifying group means. Required if
\code{multiplicative = TRUE}.}

\item{sigsq}{Numeric value specifying the variance of observations.}

\item{sigsq1, sigsq2}{Numeric value specifying the variance of observations
for each group.}

\item{sigsq_p_range}{Numeric vector specifying range of processing error
variances to consider.}

\item{sigsq_m}{Numeric value specifying the variance of measurement errors.}

\item{multiplicative}{Logical value for whether to assume multiplicative
rather than additive errors.}

\item{alpha}{Numeric value specifying type-1 error rate.}

\item{labels}{Logical value.}

\item{ylim}{Numeric vector.}
}
\value{
Plot generated by \code{\link[ggplot2]{ggplot}}.
}
\description{
Useful for choosing a sample size such that power will be adequate even if
the processing errors are larger than anticipated.
}
\examples{
# Plot ratio of variances vs. pool size with default settings
poolvar_t(sigsq = 1)

# Add processing error and other per-subject costs
poolvar_t(sigsq = 1, sigsq_p = 0.2, other_costs = 0.1)


}
