% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\docType{package}
\name{pooling-package}
\alias{pooling}
\alias{pooling-package}
\title{Fit Poolwise Regression Models}
\description{
Functions for calculating power and fitting regression models in studies
where a biomarker is measured in "pooled" samples rather than for each
individual.
}
\details{
Currently, there are two functions for estimating the covariate-adjusted
log-odds ratio relating a continuous exposure measured in pools and subject
to errors and a binary outcome. The function \code{\link{p.logreg.xerrors}}
implements homogeneous-pools logistic regression, while
\code{\link{p.dfa.xerrors}} implements the discriminant function approach.
}
