% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\docType{package}
\name{pooling}
\alias{pooling}
\alias{pooling-package}
\title{Fit Poolwise Regression Models}
\description{
Functions for calculating power and fitting regression models in studies
where a biomarker is measured in "pooled" samples rather than for each
individual.
}
\details{
\tabular{ll}{
Package: \tab pooling \cr
Type: \tab Package \cr
Version: \tab 1.1.1 \cr
Date: \tab 2018-04-05 \cr
License: \tab GPL-3 \cr
}
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen \cr \email{vandomed@gmail.com}
}
