% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolpower_t.R
\name{poolpower_t}
\alias{poolpower_t}
\title{Visualize T-test Power for Traditional vs. Pooling Study Design}
\usage{
poolpower_t(delta, var, var_pe = 0, var_me = 0, type = "two.sample",
  assay_cost = 1, other_costs = 0, pool_size = c(1, 2, 3, 10), ...)
}
\arguments{
\item{delta}{Numeric value.}

\item{var}{Numeric value specifying variance of observations.}

\item{var_pe}{Numeric value specifying variance of additive processing error.}

\item{var_me}{Numeric value specifying variance of additive measurement
error.}

\item{type}{Character string specifying type of t-test. Choices are
\code{"two.sample"}, \code{"one.sample"}, and \code{"paired"}.}

\item{assay_cost}{Numeric value specifying cost of each assay.}

\item{other_costs}{Numeric value specifying other per-subject costs.}

\item{pool_size}{Numeric vector specifying pool sizes.}

\item{...}{Arguments to pass to \code{\link[stats]{power.t.test}}.}
}
\value{
Plot of power vs. total costs generated by
\code{\link[ggplot2]{ggplot}}.
}
\description{
Useful for determining potential cost savings of a pooling study design.
}
\examples{
# Power for two-sample t-test with delta = 0.5, var = 1, and no "other" costs
# per subject
poolpower_t(delta = 0.5, var = 1)

# Repeat but for other costs per subject equal to 1/4 the assay cost
poolpower_t(delta = 0.5, var = 1, other_costs = 1/4)

# Back to no other costs, but with processing and measurement error
poolpower_t(delta = 0.5, var = 1, var_pe = 0.2, var_me = 0.1)


}
