% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolvar_t.R
\name{poolvar_t}
\alias{poolvar_t}
\title{Visualize Ratio of Variance of Each Pooled Measurement to Variance of Each
Unpooled Measurement as Function of Pool Size}
\usage{
poolvar_t(g = 1:50, var = 1, var_pe = 0, var_me = 0,
  type = "two.sample", assay_cost = 1, other_costs = 0,
  label_max = TRUE, ylim = NULL, ...)
}
\arguments{
\item{g}{Numeric vector of pool sizes to include.}

\item{var}{Numeric value specifying variance of observations.}

\item{var_pe}{Numeric value specifying variance of additive processing error.}

\item{var_me}{Numeric value specifying variance of additive measurement
error.}

\item{type}{Character string specifying type of t-test. Choices are
\code{"two.sample"}, \code{"one.sample"}, and \code{"paired"}.}

\item{assay_cost}{Numeric value specifying cost of each assay.}

\item{other_costs}{Numeric value specifying other per-subject costs.}

\item{label_max}{Logical value.}

\item{...}{Arguments to pass to \code{\link[stats]{power.t.test}}.}
}
\value{
Plot of power vs. total costs generated by
\code{\link[ggplot2]{ggplot}}.
}
\description{
Useful for determining whether pooling is a good idea, and finding the
optimal pool size if it is.
}
