% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling-data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Dataset for a Paper Under Review}
\source{
Simulated data in R.
}
\description{
Simulated data intended to mimic the motivating example from a paper under
review. Generated under GLR with true log-OR = 0.05.
}
