% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling-data.R
\docType{data}
\name{xtilde0_matched}
\alias{xtilde0_matched}
\title{Dataset with Simulated Xtilde0's for Examples in cond_logreg}
\source{
Simulated data in R.
}
\description{
Includes 50 control pools of size 1 (25 of which have replicates) and 75 of
size 2. Other variables are stored in a separate data frame called
\code{\link{pdat_matched}}.
}
