% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling-data.R
\docType{data}
\name{dat_p_linreg_yerrors}
\alias{dat_p_linreg_yerrors}
\title{Dataset for Examples in p_linreg_yerrors}
\source{
Simulated data in R.
}
\description{
List containing (1) data frame with poolwise (g, Y, X1, X2) values and (2)
list with replicate Y values.
}
