% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_dfa_xerrors2.R
\name{p_dfa_xerrors2}
\alias{p_dfa_xerrors2}
\title{Discriminant Function Approach for Estimating Odds Ratio with Gamma Exposure
Measured in Pools and Subject to Errors}
\usage{
p_dfa_xerrors2(g, y, xtilde, c = NULL, constant_or = "test",
  errors = "both", integrate_tol = 1e-08, ...)
}
\arguments{
\item{g}{Numeric vector with pool sizes, i.e. number of members in each pool.}

\item{y}{Numeric vector of poolwise \code{Y} values (number of cases in each
pool).}

\item{xtilde}{Numeric vector (or list of numeric vectors, if some pools have
replicates) with \code{Xtilde} values.}

\item{c}{Numeric matrix with poolwise \strong{\code{C}} values (if any), with
one row for each pool. Can be a vector if there is only 1 covariate.}

\item{constant_or}{Logical value for whether to assume a constant OR for
\code{X}, which means that \code{gamma_y = 0}.}

\item{errors}{Character string specifying the errors that \code{X} is subject
to. Choices are \code{"neither"}, \code{"processing"} for processing error
only, \code{"measurement"} for measurement error only, and \code{"both"}.}

\item{integrate_tol}{Numeric value specifying the \code{tol} input to
\code{\link{adaptIntegrate}}.}

\item{...}{Additional arguments to pass to \code{\link[stats]{nlminb}}.}
}
\value{
List of point estimates, objects returned by \code{\link[stats]{nlminb}},
and AICs, for one or two models depending on \code{constant_or}.
}
\description{
Assumes exposure measurements are subject to multiplicative lognormal
processing error and measurement error, and exposure given covariates and
outcome is a Gamma regression.
}
\references{
Lyles, R.H., Van Domelen, D.R., Mitchell, E.M. and Schisterman, E.F. (2015)
"A discriminant function approach to adjust for processing and measurement
error When a biomarker is assayed in pooled samples."
\emph{Int. J. Environ. Res. Public Health} \strong{12}(11): 14723--14740.

Schisterman, E.F., Vexler, A., Mumford, S.L. and Perkins, N.J. (2010) "Hybrid
pooled-unpooled design for cost-efficient measurement of biomarkers."
\emph{Stat. Med.} \strong{29}(5): 597–613.

Whitcomb, B.W., Perkins, N.J., Zhang, Z., Ye, A., and Lyles, R. H. (2012)
"Assessment of skewed exposure in case-control studies with pooling."
\emph{Stat. Med.} \strong{31}: 2461–2472.
}
