% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling-data.R
\docType{data}
\name{pdat1}
\alias{pdat1}
\title{Dataset for Examples in p_dfa_xerrors and p_logreg_xerrors}
\source{
Simulated data in R.
}
\description{
Data frame with poolwise (g, Y*, Y, Xtilde, C) values.
}
