% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling-data.R
\docType{data}
\name{pdat2}
\alias{pdat2}
\title{Dataset for Examples in p_dfa_xerrors2 and p_logreg_xerrors2}
\source{
Simulated data in R.
}
\description{
List containing (1) data frame with poolwise (g, Y, Xtilde, C) values and
(2) list of C values for members of each pool.
}
