% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\docType{package}
\name{pooling}
\alias{pooling}
\alias{pooling-package}
\title{Fit Poolwise Regression Models}
\description{
Functions for calculating power and fitting regression models in studies
where a biomarker is measured in "pooled" samples rather than for each
individual. Approaches for handling measurement error follow the framework of
Schisterman et al. (2010) <doi:10.1002/sim.3823>.
}
\details{
\tabular{ll}{
Package: \tab pooling \cr
Type: \tab Package \cr
Version: \tab 1.1.2 \cr
Date: \tab 2020-02-12 \cr
License: \tab GPL-3 \cr
}
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen \cr \email{vandomed@gmail.com}
}
